requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.13.18
botocore==1.16.18
certifi==2020.6.20
chardet==3.0.4
comtypes==1.1.7
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-embed-video==1.3.3
django-heroku==0.3.1
django-instagram==0.3.2
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
lxml==4.5.2
mysqlclient==1.4.6
Pillow==7.0.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.0.1
